// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const EstadoServicio = {
  ACTIVO: 'ACTIVO',
  CON_DEUDA: 'CON_DEUDA',
  CORTADO: 'CORTADO',
  SUSPENDIDO: 'SUSPENDIDO'
} as const

export type EstadoServicio = (typeof EstadoServicio)[keyof typeof EstadoServicio]


export const EstadoMedidor = {
  ACTIVO: 'ACTIVO',
  DANADO: 'DANADO',
  RETIRADO: 'RETIRADO',
  REEMPLAZADO: 'REEMPLAZADO'
} as const

export type EstadoMedidor = (typeof EstadoMedidor)[keyof typeof EstadoMedidor]


export const EstadoLectura = {
  REGISTRADA: 'REGISTRADA',
  CONFIRMADA: 'CONFIRMADA',
  ANULADA: 'ANULADA'
} as const

export type EstadoLectura = (typeof EstadoLectura)[keyof typeof EstadoLectura]


export const EstadoFactura = {
  PENDIENTE: 'PENDIENTE',
  PAGADA: 'PAGADA',
  VENCIDA: 'VENCIDA',
  ANULADA: 'ANULADA'
} as const

export type EstadoFactura = (typeof EstadoFactura)[keyof typeof EstadoFactura]


export const EstadoPago = {
  PENDIENTE: 'PENDIENTE',
  CONFIRMADO: 'CONFIRMADO',
  ANULADO: 'ANULADO'
} as const

export type EstadoPago = (typeof EstadoPago)[keyof typeof EstadoPago]


export const EstadoCorte = {
  PENDIENTE: 'PENDIENTE',
  EJECUTADO: 'EJECUTADO',
  CANCELADO: 'CANCELADO'
} as const

export type EstadoCorte = (typeof EstadoCorte)[keyof typeof EstadoCorte]


export const EstadoReconexion = {
  PENDIENTE: 'PENDIENTE',
  EJECUTADA: 'EJECUTADA',
  CANCELADA: 'CANCELADA'
} as const

export type EstadoReconexion = (typeof EstadoReconexion)[keyof typeof EstadoReconexion]